cropguard-ai/backend/requirements.txt
flask
flask-cors
pymongo
pillow
numpy
tensorflow
gunicorn
bcrypt
